+++ CHANGELOG.md
@@ -6,10 +6,13 @@
 * Show "repeat from" icon in repeat column
 * Show tooltip for task checkbox filters
 * Publish task calendar events to Google Calendar (#142)
+* Introduce custom scheme
+* Add note quick add field
 
 ## Bug Fixes
 
 * Fix "yearly" repeat value
+* Apply template from selected note filter
 
 # 0.9.5